Feature #1: set base url
