+++ core/network/build.gradle.kts
@@ -1,4 +1,5 @@
 import com.hankki.build_logic.setNamespace
+import java.util.Properties
 
 plugins {
     alias(libs.plugins.hankki.library)
@@ -7,6 +8,26 @@ plugins {
 
 android {
     setNamespace("core.network")
+
+    val properties = Properties().apply {
+        load(rootProject.file("local.properties").inputStream())
+    }
+
+    buildTypes{
+        debug {
+            val devUrl = properties["hankkiDevUrl"] as? String ?: ""
+            buildConfigField("String", "BASE_URL", devUrl)
+        }
+
+        release {
+            val prodUrl = properties["hankkiProdUrl"] as? String ?: ""
+            buildConfigField("String", "BASE_URL", prodUrl)
+        }
+    }
+
+    buildFeatures {
+        buildConfig = true
+    }
 }
 
 dependencies{
